Add files via upload
This 3-D models galery is still eork in progress. I am tinkering on the best expression. The hydro moniker refers to the use of fluid dynamics computation, since the eddies resamble fibonacci's spirals governed by phi. Hurst, Strouhal and von Karman are deployed in this particular model. Thanks to the generous suppoert by AWS program for startups, These models will be tested in real-life and the results shall be published. The calculations of the lattices do take some time, nevertheless.